@@ -5,7 +5,7 @@ homepage: https://github.com/devdennysegura/flutter-auth0
 version: 0.6.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=2.0.0 <3.0.0"
 
 dependencies:
