@@ -7,13 +7,12 @@ class Auth0Auth {
   Auth0Auth._(this.client, this.clientId);
 
   factory Auth0Auth(String clientId, String url, {dynamic bearer}) {
-    assert(clientId != null);
     final _client = new Auth0Client(url, telemetry: telemetry, token: bearer);
     return Auth0Auth._(_client, clientId);
   }
 
-  Future<dynamic> responseHandler(http.Response response) {
-    if (response.statusCode == 200) {
+  Future<dynamic>? responseHandler(http.Response? response) {
+    if (response!.statusCode == 200) {
       return jsonDecode(response.body);
     }
     else if (response.statusCode == 401) {
@@ -102,8 +101,8 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'authorization_code',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
     } catch (e, stackTrace) {
       throw new Auth0Exeption(description: e, stackTrace: stackTrace);
     }
@@ -130,9 +129,9 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'http://auth0.com/oauth/grant-type/password-realm',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
-    } catch (e, stackTrace) {
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
+    } on (e, stackTrace) {
       throw new Auth0Exeption(
           name: e['name'] ?? e['error'],
           description:
@@ -158,8 +157,8 @@ class Auth0Auth {
           'client_id': this.clientId,
           'grant_type': 'refresh_token',
         });
-      http.Response res = await this.client.mutate('/oauth/token', payload);
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.mutate('/oauth/token', payload);
+      return await responseDataHandler(res!);
     } catch (e, stackTrace) {
       throw new Auth0Exeption(description: e, stackTrace: stackTrace);
     }
@@ -172,8 +171,8 @@ class Auth0Auth {
   //
   Future<dynamic> getUserInfo() async {
     try {
-      http.Response res = await this.client.query('/userinfo');
-      return await responseDataHandler(res);
+      http.Response? res = await this.client.query('/userinfo');
+      return await responseDataHandler(res!);
     } catch (e, stackTrace) {
       throw new Auth0Exeption(description: e, stackTrace: stackTrace);
     }
@@ -195,8 +194,8 @@ class Auth0Auth {
     return this
         .client
         .mutate('/oauth/revoke', payload)
-        .then((http.Response response) {
-      if (response.statusCode == 200) {
+        .then((http.Response? response) {
+      if (response!.statusCode == 200) {
         return {};
       }
       throw new Auth0Exeption(description: jsonDecode(response.body));
@@ -229,8 +228,8 @@ class Auth0Auth {
     return this
         .client
         .mutate('/dbconnections/change_password', payload)
-        .then((http.Response response) {
-      if (response.statusCode == 200) {
+        .then((http.Response? response) {
+      if (response!.statusCode == 200) {
         return true;
       }
       throw jsonDecode(response.body);
@@ -254,11 +253,11 @@ class Auth0Auth {
       var payload = Map.from(params)..addAll({'client_id': this.clientId});
       if (params['metadata'] != null)
         payload..addAll({'user_metadata': params['metadata']});
-      http.Response res = await this.client.mutate(
+      http.Response? res = await this.client.mutate(
             '/dbconnections/signup',
             payload,
           );
-      return await responseDataHandler(res);
+      return await responseDataHandler(res!);
     } catch (e, stackTrace) {
       throw new Auth0Exeption(
           name: e['name'], description: e['message'] ?? e['description'], stackTrace: stackTrace);
